backend/api/requirements.txt
absl-py==0.9.0
astunparse==1.6.3
azure-cognitiveservices-speech==1.14.0
azure-core==1.9.0
azure-storage-blob==12.6.0
backcall==0.1.0
boto==2.49.0
boto3==1.17.76
botocore==1.20.76
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
Flask==2.0.0
fvcore==0.1.3.post20210317
gast==0.3.3
google-api-core==1.27.0
google-auth==1.30.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
graphviz==0.14
grpcio==1.29.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
imutils==0.5.3
install==1.3.4
iopath==0.1.6
ipykernel==5.4.3
ipython==7.15.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.15.1
jupyter-client==6.1.11
jupyter-core==4.7.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.2.1
msrest==0.6.19
mysql-connector-python==8.0.25
networkx==2.4
numpy==1.18.4
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.9
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
portalocker==2.2.1
prompt-toolkit==3.0.5
protobuf==3.12.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch3d==0.4.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.4.2
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.0.0
tifffile==2020.5.30
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.2.3
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==1.3.3
yacs==0.1.8
zipp==3.1.0
